added liek 27 thousand different typecasts on the mult function because we need to make sure it sign extends properly but im not too sure if itll work we'll see... does Dr. David Thomas actually read our commiits if so helloo!!!

+++ instructions.cpp
@@ -76,8 +76,8 @@ void mtlo(uint8_t rs){
 
 //overflow condition ?
 void mult(uint8_t rs, uint8_t rt){
-	Register[LO] = (uint32_t)(((((uint64_t)Register[rs])*((uint64_t)Register[rt]))) & 0xFFFFFFFF);
-	Register[HI] = (uint32_t)(((((uint64_t)Register[rs])*((uint64_t)Register[rt])) >> 32) & 0xFFFFFFFF);
+	Register[LO] = (uint32_t)(((((int64_t)((int32_t)Register[rs]))*(int64_t)(((int32_t)Register[rt])))) & 0xFFFFFFFF);
+	Register[HI] = (uint32_t)(((((int64_t)((int32_t)Register[rs]))*((int64_t)((int32_t)Register[rt]))) >> 32) & 0xFFFFFFFF);
     PC_advance(default_advance);
 }
